backend/requirements.txt
fastapi
uvicorn[standard]
httpx
apscheduler
onnxruntime
numpy
pandas
scikit-learn
joblib
groq
python-dotenv
websockets